// projects/wa-motif-open-api/catalog-service/src/lib/model/serviceOperationProperties.ts
/**
 * Motif Service Catalog API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: info@vipera.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ServiceOperationProperties { 
    description: string;
    pluginName: string;
    inputParams?: string;
    outputParams?: string;
    counted: boolean;
    secure: boolean;
    sessionless: boolean;
    encryptActive: boolean;
}